requirements.txt
pytorch-lightning==1.5.6
PyYAML==6.0
tensorboard>=2.2.0
torch==1.11.0
torchsummary==1.5.1
torchvision==0.12.0
opencv-python~=4.5.5.64
numpy~=1.21.6
Pillow~=9.1.0
scipy
librosa
soundfile